DMMD_metric/requirements.txt
torch
numpy
Pillow
tqdm
transformers
accelerate
absl-py